bugfixes + undisclosed experiments with OD filter and normalization

+++ data_stack.py
@@ -900,18 +900,18 @@ class data:
     def crop_registed_images(self, images, min_xshift, max_xshift, min_yshift, max_yshift):
                 
         # if the image is moved to the right (positive) we need to crop the left side 
-        xleft = np.ceil(max_xshift)
+        xleft = int(np.ceil(max_xshift))
         if xleft < 0:
             xleft = 0
         # if the image is moved to the left (negative) we need to crop the right side 
-        xright = np.floor(self.n_cols+min_xshift)
+        xright = int(np.floor(self.n_cols+min_xshift))
         if xright>(self.n_cols):
             xright = self.n_cols
         
-        ybottom = np.ceil(max_yshift)
+        ybottom = int(np.ceil(max_yshift))
         if ybottom <0:
             ybottom = 0
-        ytop = np.floor(self.n_rows+min_yshift)
+        ytop = int(np.floor(self.n_rows+min_yshift))
         if ytop > (self.n_rows):
             ytop = self.n_rows
